added config support

+++ src/actions/removeProj.ts
@@ -8,24 +8,30 @@ const STORAGE_PATH = homedir() + "/.tasklist/tasklist.json";
 let tasks = JSON.parse(fs.readFileSync(STORAGE_PATH, "utf-8"));
 
 // function to remove a project
-export default async function removeProj(proj: string) {
+export default async function removeProj(config: configType, proj: string) {
   if (Object.keys(tasks).includes(proj)) {
-    delete tasks[proj];
-    console.clear();
-    const confirmDelete = await inquirer.prompt({
-      name: "choice",
-      type: "confirm",
-      prefix: "❓",
-      message: `Are you sure to delete "${proj}"?`,
-      default() {
-        return false;
-      },
-    });
-    if (confirmDelete.choice) {
+    if ((config?.default ?? "default") !== proj) {
+      delete tasks[proj];
+      console.clear();
+      const confirmDelete = await inquirer.prompt({
+        name: "choice",
+        type: "confirm",
+        prefix: config?.emoji ?? true ? "❓" : undefined,
+        message: `Are you sure to delete "${proj}"?`,
+        default() {
+          return false;
+        },
+      });
+      if (confirmDelete.choice) {
+        console.log(
+          `Project ${chalk.yellow(proj)} has been ${chalk.red("deleted.")}`
+        );
+        fs.writeFileSync(STORAGE_PATH, JSON.stringify(tasks));
+      }
+    } else {
       console.log(
-        `Project ${chalk.yellow(proj)} has been ${chalk.red("deleted.")}`
+        `You cannot delete the default project ${chalk.yellow(proj)}.`
       );
-      fs.writeFileSync(STORAGE_PATH, JSON.stringify(tasks));
     }
   } else {
     console.log(`${chalk.red(proj)}: No such project.`);
